[REFACTOR] rename buffer file name
