+++ tests/test_buffer.py
@@ -1,4 +1,4 @@
-from buffer import BufferManager
+from command_buffer import BufferManager
 
 
 def test_bufferManager_flow():
